Demo/paddle_ocr/requirements.txt
aiohttp==3.11.12
fastapi==0.115.12
numpy==1.23.0
paddleocr==2.9.1
paddleocr.egg==info
Pillow==11.1.0
python-dotenv==1.0.1
uvicorn==0.34.0
